pyproject.toml
[project]
name = "modelscope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.2.0",
    "openai",
]
